Sellapp python documentation
Normal sellapp embeds are quite ugly so with this we have our own SDK python files where we can customize whatever we want in order to change the way embeds work. So instead of some random blue 1999 looking page popup, we can have really cool ones with animations in the back.