Code cleanup for sellapp embeds
So with this commit essentially this is a test of how new sellapp embeds will look like. The old ones look pretty bad its just a basic font and solid blue boxes. But with this one, there are custom scroll bars as well as a page where it actually fits our color scheme instead of some ugly and random box that says buy now in some 1999 font.